dashboard

+++ src/app/pages/clients/client-profile/client-profile.component.ts
@@ -14,10 +14,12 @@ import { AddDeviceComponent } from './add-device/add-device.component';
 import { AddLocationComponent } from './add-location/add-location.component';
 import { AddStoreComponent } from './add-store/add-store.component';
 import { map } from 'rxjs/operators';
+import { ConfirmationService, Message } from 'primeng/api';
+
 @Component({
   templateUrl: './client-profile.component.html',
   styleUrls: ['./client-profile.component.scss'],
-  providers: [MessageService, DialogService]
+  providers: [MessageService, DialogService, ConfirmationService]
 })
 export class ClientProfileComponent implements OnInit {
   storeList$: Observable<Store[] | any> = of([]);
@@ -32,7 +34,7 @@ export class ClientProfileComponent implements OnInit {
   filiterStore: Store[] = [];
   filiterDevice: Device[] = [];
   filiterBranches: Branch[] = []
-  constructor(private readonly _messageService: MessageService, private _activatedRoute: ActivatedRoute, private _cdr: ChangeDetectorRef, public dialogService: DialogService, public messageService: MessageService, private _apiService: ApiService, private _translate: TranslateService) {
+  constructor(private confirmationService: ConfirmationService, private readonly _messageService: MessageService, private _activatedRoute: ActivatedRoute, private _cdr: ChangeDetectorRef, public dialogService: DialogService, public messageService: MessageService, private _apiService: ApiService, private _translate: TranslateService) {
     this._activatedRoute.queryParams.subscribe(params => {
       this.clientId = +params['id'];
     });
@@ -241,32 +243,65 @@ export class ClientProfileComponent implements OnInit {
 
 
   deleteStore = (data: Store): void => {
-    this._apiService.apiName = `clients/${this.clientId}/stores/`;
-    this._apiService.delete(data.id).subscribe((response) => {
-      if (response.error_message == 'success') {
-        this.loadAllStore();
-        this._cdr.detectChanges();
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to delete this record?',
+      icon: 'pi pi-info-circle',
+      acceptLabel: 'Yes',
+      rejectLabel: 'No',
+      accept: () => {
+        this._apiService.apiName = `clients/${this.clientId}/stores/`;
+        this._apiService.delete(data.id).subscribe((response) => {
+          if (response.error_message == 'success') {
+            this.loadAllStore();
+            this._cdr.detectChanges();
+          }
+        });
+      },
+      reject: () => {
+        return;
       }
-    });
+    })
   }
 
   deleteBranch = (data: Branch): void => {
-    this._apiService.apiName = `clients/${this.clientId}/branches/`;
-    this._apiService.delete(data.id).subscribe((response) => {
-      if (response.error_message == 'success') {
-        this.loadAllBranch();
-        this._cdr.detectChanges();
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to delete this record?',
+      icon: 'pi pi-info-circle',
+      acceptLabel: ' Yes',
+      rejectLabel: 'No',
+      accept: () => {
+        this._apiService.apiName = `clients/${this.clientId}/branches/`;
+        this._apiService.delete(data.id).subscribe((response) => {
+          if (response.error_message == 'success') {
+            this.loadAllBranch();
+            this._cdr.detectChanges();
+          }
+        });
+      },
+      reject: () => {
+        return;
       }
-    });
+    })
   }
 
   deleteDevice = (data: Device): void => {
-    this._apiService.apiName = `clients/${this.clientId}/devices/`;
-    this._apiService.delete(data.id).subscribe((response) => {
-      if (response.error_message == 'success') {
-        this.loadAllDevice();
-        this._cdr.detectChanges();
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to delete this record?',
+      icon: 'pi pi-info-circle',
+      acceptLabel: 'Yes',
+      rejectLabel: 'No',
+      accept: () => {
+        this._apiService.apiName = `clients/${this.clientId}/devices/`;
+        this._apiService.delete(data.id).subscribe((response) => {
+          if (response.error_message == 'success') {
+            this.loadAllDevice();
+            this._cdr.detectChanges();
+          }
+        });
+      },
+      reject: () => {
+        return;
       }
-    });
+    })
   }
 }
